refactor node assignment to be a method on skyhookNodes and use floor for capacity calculations
