+++ operator/internal/controller/cluster_state_v2_test.go
@@ -466,4 +466,446 @@ var _ = Describe("CleanupRemovedNodes", func() {
 			Expect(result).To(BeFalse())
 		})
 	})
+
+	Describe("AssignNodeToCompartment", func() {
+		It("should assign node to compartment", func() {
+			compartment := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment",
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label": "test-value"},
+				},
+			})
+
+			node := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node",
+					Labels: map[string]string{"test-label": "test-value"},
+				},
+			}
+
+			skyhook := &v1alpha1.Skyhook{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test-skyhook",
+				},
+			}
+
+			skyhookNode, err := wrapper.NewSkyhookNode(node, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+
+			skyhookNodes := &skyhookNodes{
+				skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+				nodes:        []wrapper.SkyhookNode{skyhookNode},
+				compartments: make(map[string]*wrapper.Compartment),
+			}
+			skyhookNodes.AddCompartment("test-compartment", compartment)
+
+			compartmentName, err := skyhookNodes.AssignNodeToCompartment(skyhookNode)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment"))
+		})
+
+		It("should assign node to default compartment", func() {
+			compartment := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment",
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label": "test-value"},
+				},
+			})
+
+			defaultCompartment := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: v1alpha1.DefaultCompartmentName,
+			})
+
+			node := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node",
+					Labels: map[string]string{"test-label": "test-value-other"},
+				},
+			}
+
+			skyhook := &v1alpha1.Skyhook{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test-skyhook",
+				},
+			}
+
+			skyhookNode, err := wrapper.NewSkyhookNode(node, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+
+			skyhookNodes := &skyhookNodes{
+				skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+				nodes:        []wrapper.SkyhookNode{skyhookNode},
+				compartments: make(map[string]*wrapper.Compartment),
+			}
+			skyhookNodes.AddCompartment("test-compartment", compartment)
+			skyhookNodes.AddCompartment(v1alpha1.DefaultCompartmentName, defaultCompartment)
+
+			compartmentName, err := skyhookNodes.AssignNodeToCompartment(skyhookNode)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal(v1alpha1.DefaultCompartmentName))
+		})
+
+		It("should assign node to compartment with safer strategy", func() {
+			compartment1 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-1",
+				Strategy: &v1alpha1.DeploymentStrategy{
+					Fixed: &v1alpha1.FixedStrategy{
+						InitialBatch: ptr(1),
+					},
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-1": "test-value-1"},
+				},
+			})
+
+			compartment2 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-2",
+				Strategy: &v1alpha1.DeploymentStrategy{
+					Linear: &v1alpha1.LinearStrategy{
+						InitialBatch: ptr(1),
+					},
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-2": "test-value-2"},
+				},
+			})
+
+			compartment3 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-3",
+				Strategy: &v1alpha1.DeploymentStrategy{
+					Exponential: &v1alpha1.ExponentialStrategy{
+						InitialBatch: ptr(1),
+					},
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-3": "test-value-3"},
+				},
+			})
+
+			fixedNode := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node-fixed",
+					Labels: map[string]string{"test-label-1": "test-value-1", "test-label-2": "test-value-2", "test-label-3": "test-value-3"},
+				},
+			}
+
+			linearNode := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node-linear",
+					Labels: map[string]string{"test-label-2": "test-value-2", "test-label-3": "test-value-3"},
+				},
+			}
+
+			exponentialNode := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node-exponential",
+					Labels: map[string]string{"test-label-3": "test-value-3"},
+				},
+			}
+
+			skyhook := &v1alpha1.Skyhook{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test-skyhook",
+				},
+			}
+
+			skyhookNodeFixed, err := wrapper.NewSkyhookNode(fixedNode, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+			skyhookNodeLinear, err := wrapper.NewSkyhookNode(linearNode, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+			skyhookNodeExponential, err := wrapper.NewSkyhookNode(exponentialNode, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+
+			skyhookNodes := &skyhookNodes{
+				skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+				nodes:        []wrapper.SkyhookNode{skyhookNodeFixed, skyhookNodeLinear, skyhookNodeExponential},
+				compartments: make(map[string]*wrapper.Compartment),
+			}
+			skyhookNodes.AddCompartment("test-compartment-1", compartment1)
+			skyhookNodes.AddCompartment("test-compartment-2", compartment2)
+			skyhookNodes.AddCompartment("test-compartment-3", compartment3)
+
+			compartmentName, err := skyhookNodes.AssignNodeToCompartment(skyhookNodeFixed)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment-1"))
+			compartmentName, err = skyhookNodes.AssignNodeToCompartment(skyhookNodeLinear)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment-2"))
+			compartmentName, err = skyhookNodes.AssignNodeToCompartment(skyhookNodeExponential)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment-3"))
+		})
+
+		It("should assign node to compartment with smaller count budget", func() {
+			compartment1 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-1",
+				Budget: v1alpha1.DeploymentBudget{
+					Count: ptr(1),
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-1": "test-value-1"},
+				},
+			})
+
+			compartment2 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-2",
+				Budget: v1alpha1.DeploymentBudget{
+					Count: ptr(2),
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-2": "test-value-2"},
+				},
+			})
+
+			node1 := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node-1",
+					Labels: map[string]string{"test-label-1": "test-value-1", "test-label-2": "test-value-2"},
+				},
+			}
+
+			skyhook := &v1alpha1.Skyhook{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test-skyhook-1",
+				},
+			}
+
+			skyhookNode1, err := wrapper.NewSkyhookNode(node1, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+
+			skyhookNodes := &skyhookNodes{
+				skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+				nodes:        []wrapper.SkyhookNode{skyhookNode1},
+				compartments: make(map[string]*wrapper.Compartment),
+			}
+			skyhookNodes.AddCompartment("test-compartment-1", compartment1)
+			skyhookNodes.AddCompartment("test-compartment-2", compartment2)
+
+			compartmentName, err := skyhookNodes.AssignNodeToCompartment(skyhookNode1)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment-1"))
+		})
+
+		It("should assign node to compartment with smaller percent budget", func() {
+			compartment1 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-1",
+				Budget: v1alpha1.DeploymentBudget{
+					Percent: ptr(10),
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-1": "test-value-1"},
+				},
+			})
+
+			compartment2 := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-2",
+				Budget: v1alpha1.DeploymentBudget{
+					Percent: ptr(20),
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-2": "test-value-2"},
+				},
+			})
+
+			node1 := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node-1",
+					Labels: map[string]string{"test-label-1": "test-value-1", "test-label-2": "test-value-2"},
+				},
+			}
+
+			skyhook := &v1alpha1.Skyhook{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "test-skyhook-1",
+				},
+			}
+
+			skyhookNode1, err := wrapper.NewSkyhookNode(node1, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+
+			skyhookNodes := &skyhookNodes{
+				skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+				nodes:        []wrapper.SkyhookNode{skyhookNode1},
+				compartments: make(map[string]*wrapper.Compartment),
+			}
+			skyhookNodes.AddCompartment("test-compartment-1", compartment1)
+			skyhookNodes.AddCompartment("test-compartment-2", compartment2)
+
+			compartmentName, err := skyhookNodes.AssignNodeToCompartment(skyhookNode1)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment-1"))
+		})
+
+		It("should assign to compartment with higher percent but smaller effective capacity due to fewer matching nodes", func() {
+			// Test that a compartment with a higher percent but fewer matching nodes
+			// can have a smaller effective capacity and win the assignment
+
+			// Compartment A: 50% budget, matches 10 nodes total
+			compartmentA := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-1",
+				Budget: v1alpha1.DeploymentBudget{
+					Percent: ptr(50), // 50% of 10 = 5 capacity
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-1": "test-value-1"},
+				},
+				Strategy: &v1alpha1.DeploymentStrategy{Fixed: &v1alpha1.FixedStrategy{}},
+			})
+
+			// Compartment B: 80% budget, matches only 2 nodes total
+			compartmentB := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+				Name: "test-compartment-2",
+				Budget: v1alpha1.DeploymentBudget{
+					Percent: ptr(80), // 80% of 2 = floor(1.6) = 1 capacity
+				},
+				Selector: metav1.LabelSelector{
+					MatchLabels: map[string]string{"test-label-2": "test-value-2"},
+				},
+				Strategy: &v1alpha1.DeploymentStrategy{Fixed: &v1alpha1.FixedStrategy{}},
+			})
+
+			// Target node matches both compartments
+			targetNode := &corev1.Node{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   "test-node-1",
+					Labels: map[string]string{"test-label-1": "test-value-1", "test-label-2": "test-value-2"},
+				},
+			}
+
+			// Create all nodes
+			allNodesList := []*corev1.Node{
+				targetNode,
+				// 9 more nodes that match compartment A (test-label-1=test-value-1) but not B
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a1", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a2", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a3", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a4", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a5", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a6", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a7", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a8", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-a9", Labels: map[string]string{"test-label-1": "test-value-1"}}},
+				// 1 more node that matches compartment B (test-label-2=test-value-2) but not A
+				{ObjectMeta: metav1.ObjectMeta{Name: "node-b1", Labels: map[string]string{"test-label-2": "test-value-2"}}},
+			}
+
+			skyhook := &v1alpha1.Skyhook{ObjectMeta: metav1.ObjectMeta{Name: "test-skyhook"}}
+			var allNodes []wrapper.SkyhookNode
+			for _, n := range allNodesList {
+				sn, err := wrapper.NewSkyhookNode(n, skyhook)
+				Expect(err).NotTo(HaveOccurred())
+				allNodes = append(allNodes, sn)
+			}
+
+			targetSkyhookNode, err := wrapper.NewSkyhookNode(targetNode, skyhook)
+			Expect(err).NotTo(HaveOccurred())
+
+			skyhookNodes := &skyhookNodes{
+				skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+				nodes:        allNodes,
+				compartments: make(map[string]*wrapper.Compartment),
+			}
+			skyhookNodes.AddCompartment("test-compartment-1", compartmentA)
+			skyhookNodes.AddCompartment("test-compartment-2", compartmentB)
+
+			// Should assign to "test-compartment-2" because:
+			// - test-compartment-1: 50% × 10 nodes = 5 capacity
+			// - test-compartment-2: 80% × 2 nodes = 1 capacity (smaller wins)
+			compartmentName, err := skyhookNodes.AssignNodeToCompartment(targetSkyhookNode)
+			Expect(err).NotTo(HaveOccurred())
+			Expect(compartmentName).To(Equal("test-compartment-2"), "higher percent but fewer matching nodes = smaller capacity")
+		})
+
+		It("should use lexicographic order as final tiebreaker when strategy and capacity are identical", func() {
+			testCases := []struct {
+				name           string
+				compartmentA   string
+				compartmentB   string
+				selectorA      map[string]string
+				selectorB      map[string]string
+				expectedWinner string
+			}{
+				{
+					name:           "zebra vs apple - apple wins",
+					compartmentA:   "zebra-compartment",
+					compartmentB:   "apple-compartment",
+					selectorA:      map[string]string{"env": "prod"},
+					selectorB:      map[string]string{"tier": "frontend"},
+					expectedWinner: "apple-compartment",
+				},
+				{
+					name:           "production vs development - development wins",
+					compartmentA:   "production",
+					compartmentB:   "development",
+					selectorA:      map[string]string{"env": "prod"},
+					selectorB:      map[string]string{"tier": "frontend"},
+					expectedWinner: "development",
+				},
+				{
+					name:           "comp-2 vs comp-1 - comp-1 wins",
+					compartmentA:   "comp-2",
+					compartmentB:   "comp-1",
+					selectorA:      map[string]string{"env": "prod"},
+					selectorB:      map[string]string{"tier": "frontend"},
+					expectedWinner: "comp-1",
+				},
+			}
+
+			for _, tc := range testCases {
+				By(tc.name)
+
+				node := &corev1.Node{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:   "test-node",
+						Labels: map[string]string{"env": "prod", "tier": "frontend", "zone": "us-west"},
+					},
+				}
+
+				skyhook := &v1alpha1.Skyhook{ObjectMeta: metav1.ObjectMeta{Name: "test-skyhook"}}
+				skyhookNode, err := wrapper.NewSkyhookNode(node, skyhook)
+				Expect(err).NotTo(HaveOccurred())
+
+				compartmentA := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+					Name: tc.compartmentA,
+					Budget: v1alpha1.DeploymentBudget{
+						Count: ptr(5),
+					},
+					Selector: metav1.LabelSelector{
+						MatchLabels: tc.selectorA,
+					},
+					Strategy: &v1alpha1.DeploymentStrategy{Fixed: &v1alpha1.FixedStrategy{}},
+				})
+
+				compartmentB := wrapper.NewCompartmentWrapper(&v1alpha1.Compartment{
+					Name: tc.compartmentB,
+					Budget: v1alpha1.DeploymentBudget{
+						Count: ptr(5), // Same capacity
+					},
+					Selector: metav1.LabelSelector{
+						MatchLabels: tc.selectorB,
+					},
+					Strategy: &v1alpha1.DeploymentStrategy{Fixed: &v1alpha1.FixedStrategy{}}, // Same strategy
+				})
+
+				skyhookNodes := &skyhookNodes{
+					skyhook:      wrapper.NewSkyhookWrapper(skyhook),
+					nodes:        []wrapper.SkyhookNode{skyhookNode},
+					compartments: make(map[string]*wrapper.Compartment),
+				}
+				skyhookNodes.AddCompartment(tc.compartmentA, compartmentA)
+				skyhookNodes.AddCompartment(tc.compartmentB, compartmentB)
+
+				compartmentName, err := skyhookNodes.AssignNodeToCompartment(skyhookNode)
+				Expect(err).NotTo(HaveOccurred())
+				Expect(compartmentName).To(Equal(tc.expectedWinner), "lexicographic tie-break")
+
+				// Run multiple times to ensure consistency
+				for i := 0; i < 5; i++ {
+					result, err := skyhookNodes.AssignNodeToCompartment(skyhookNode)
+					Expect(err).NotTo(HaveOccurred())
+					Expect(result).To(Equal(compartmentName), "should be deterministic across multiple calls")
+				}
+			}
+		})
+	})
 })
